requirements.txt
gunicorn
Django
django-heroku
djangorestframework
django-cors-headers
opencv-python
tensorflow-cpu